Conflicts:
	applications/plugins/image/CollisionToCarvingEngine.h
	examples/Demos/SkullOperations/Test.scn

Former-commit-id: 7fa9e30556b126727c96047ed4f6c0526ce17d50

+++ modules/SofaLoader/MeshVTKLoader.cpp
@@ -302,7 +302,7 @@ bool MeshVTKLoader::setInputsMesh()
 
             if (cellData == NULL) return false;
 
-            cellData->resize(numSubPolyLines.size());
+            cellData->resize((int)numSubPolyLines.size());
 
             for (size_t ii = 0;  ii < numSubPolyLines.size(); ii++)
                 cellData->data[ii] = numSubPolyLines[ii];
